C:Mount - Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@45415 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/c/Mount.c
@@ -1936,7 +1936,7 @@ void ShowFault(LONG code, char *s, ...)
 	l += 2;
 	Fault(code, NULL, &buf[l], sizeof(buf) - l);
 	if (buf[l] == 0)
-	    snprintf(&buf[l], sizeof(buf) - l, "%d", code);
+	    snprintf(&buf[l], sizeof(buf) - l, "%ld", (long)code);
 	buf[sizeof(buf)-1] = 0;
 	if (IsCli)
 	{
